added memory location struct to prepare drafting out memory ssa pass
